Allow assignment variable "`:=`" within queries
The assignment operator "`:=`" is intepreted as an invalid named variable by yesql, causing a parse error (see issue #80).  Yesql already has special allowance in the statement.bnf file for the casting operator "`::`".  I've extended that support to also include the assignment operator "`:=`".

This has a side effect of intepreting named-variables beginning with an "`=`" sign (valid clojure) as assignments to the rest of the variable name. For example:

```clojure
"SELECT * FROM country WHERE country = :=invalid-named-variable;"
=> ["SELECT * FROM country WHERE country = :=invalid-named-variable;"]
```

It does not throw a parse error because the named-variable `=invalid-named-variable` is interpreted as an assignment `:= invalid-named-variable`.